+++ storage.go
@@ -5,12 +5,14 @@ import (
 	"time"
 )
 
+// MemoryStore storage.
 type MemoryStore struct {
 	token        string
 	refreshToken string
 	expiresTime  time.Time
 }
 
+// Token returns valid token.
 func (s *MemoryStore) Token() string {
 	if time.Now().Before(s.expiresTime) {
 		return s.token
@@ -18,6 +20,7 @@ func (s *MemoryStore) Token() string {
 	return ""
 }
 
+// Refresh token
 func (s *MemoryStore) Refresh(c *Client) (err error) {
 	r := &TokenRequest{}
 	if s.refreshToken != "" {