+++ CHANGELOG.md
@@ -2,12 +2,33 @@
 ## [Unreleased]
 
 
+<a name="v0.2.0"></a>
+## [v0.2.0] - 2020-12-30
+### Features
+- compensation mechanism for refresh token
+
+
+<a name="v0.2.0-beta.1"></a>
+## [v0.2.0-beta.1] - 2020-12-29
+### Features
+- ExponentialBackOff support for API.
+
+
+<a name="v0.1.3"></a>
+## [v0.1.3] - 2020-12-29
+### Bug Fixes
+- token data race bug
+
+### Features
+- Query the list of home that the user lives in
+
+
 <a name="v0.1.2"></a>
-## [v0.1.2] - 2020-09-08
+## [v0.1.2] - 2020-12-07
 ### Code Refactoring
+- Make sure request struct implements `RequestBody` interface.
 - Some cases HTTP DELETE need HTTP body
 - Refactor pairing
-- Make sure request struct implements `RequestBody` interface.
 
 ### Features
 - Smart Home Management
@@ -41,6 +62,9 @@
 - More powerful API Client
 
 
-[Unreleased]: https://github.com/yunjuiot/tuyacloud/compare/v0.1.2...HEAD
+[Unreleased]: https://github.com/yunjuiot/tuyacloud/compare/v0.2.0...HEAD
+[v0.2.0]: https://github.com/yunjuiot/tuyacloud/compare/v0.2.0-beta.1...v0.2.0
+[v0.2.0-beta.1]: https://github.com/yunjuiot/tuyacloud/compare/v0.1.3...v0.2.0-beta.1
+[v0.1.3]: https://github.com/yunjuiot/tuyacloud/compare/v0.1.2...v0.1.3
 [v0.1.2]: https://github.com/yunjuiot/tuyacloud/compare/v0.1.1...v0.1.2
 [v0.1.1]: https://github.com/yunjuiot/tuyacloud/compare/v0.1.0...v0.1.1